Update kysmet.v
Threw in a few more things. The decode so far just straight up copies the opcode into the new opcode value, though I have visions of it also having outputs that determine things like if we need a no-op for the next instruction, if it is a jumpf and the PC needs to not increment, etc.
Most things like that are dependent on single commands, so by sending flags with the command itself, we could easily be able to tell.